requirements.txt
PyQt5>=5.15.0
pandas>=1.3.0
numpy>=1.21.0
xlsxwriter>=3.0.0
aiohttp>=3.8.0
python-dateutil>=2.8.2
requests>=2.26.0
Pillow>=10.0.0
pyinstaller>=6.0.0 